권한확인 후 api

+++ example-app/app/Http/Controllers/ArticleController.php
@@ -11,13 +11,15 @@ class ArticleController extends Controller
     public function __construct()
     {
         $this->middleware('auth')
-        ->except(['index', 'show']);
+            ->except(['index', 'show']);
     }
-    public function create() {
+    public function create()
+    {
         return view('articles/create');
     }
 
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         // 비어있지않고, 문자열이고, 255자를 넘으면 안된다.
         $input = $request->validate([
             'body' => [
@@ -51,7 +53,7 @@ class ArticleController extends Controller
 
         // // DB 파사드를 이용하는 방법
         // DB::statement("INSERT INTO articles (body, user_id) VALUES (:body, :userId)", ['body' => $input['body'], 'userId' => Auth::id()]);
-        
+
         // // 쿼리 빌더를 사용하는 방법
         // DB::table('articles')->insert([
         //     'body' => $input['body'],
@@ -78,35 +80,36 @@ class ArticleController extends Controller
         return redirect()->route('articles.index');
     }
 
-    public function index(Request $request) {
+    public function index(Request $request)
+    {
         // $page = $request->input('page', 1);
         $perPage = $request->input('per_page', 3);
         // $skip = ($page -1) * $perPage;
-        
+
         $articles = Article::with('user')
             ->select('id', 'body', 'user_id', 'created_at')
             // ->skip($skip)
             // ->take($perPage)
             ->latest()
             ->paginate($perPage);
-            // ->orderby('created_at', 'desc')
-            // ->oldest()
-            // ->orderby('body', 'asc')
-            // ->get();
-            // $articles->withQueryString();
-            // $articles->appends(['filter' => 'name']);
-            // $totalCount = Article::count();
-            // $now = Carbon::now();
-            // $past = clone $now;
-            // $past->subHours(3);
-            // dd($now->diffInMinutes($past));
+        // ->orderby('created_at', 'desc')
+        // ->oldest()
+        // ->orderby('body', 'asc')
+        // ->get();
+        // $articles->withQueryString();
+        // $articles->appends(['filter' => 'name']);
+        // $totalCount = Article::count();
+        // $now = Carbon::now();
+        // $past = clone $now;
+        // $past->subHours(3);
+        // dd($now->diffInMinutes($past));
         // $results = DB::table('articles as a')
         //     ->join('users as u', 'a.user_id', '=', 'u.id')
         //     ->select(['a.*', 'u.name'])
         //     ->latest()
         //     ->paginate($perPage);
         return view(
-            'articles.index', 
+            'articles.index',
             [
                 'articles' => $articles,
                 // 'results' => $results
@@ -120,19 +123,31 @@ class ArticleController extends Controller
 
     // Route::get('articles/{id}', function($id) {
     //     $article = Article::find($id);
-        
+
     //     return view('articles.show', ['article' => $article]);
     // });
     // 위를 라우트 모델 바인딩을 적용하면
-    public function show(Article $article) {
+    public function show(Article $article)
+    {
         return view('articles.show', ['article' => $article]);
     }
 
-    public function edit(Article $article) {
+    public function edit(Article $article)
+    {
+        // 컨트롤러 헬퍼를 사용
+        $this->authorize('edit', $article);
         return view('articles.edit', ['article' => $article]);
     }
 
-    public function update(Article $article, Request $request) {
+    public function update(Article $article, Request $request)
+    {
+        // if (!Auth::user()->can('update', $article)) {
+        //     abort(403);
+        // };
+
+        // 컨트롤러 헬퍼를 사용
+        $this->authorize('update', $article);
+
         $input = $request->validate([
             'body' => [
                 'required',
@@ -140,14 +155,18 @@ class ArticleController extends Controller
                 'max:255'
             ],
         ]);
-    
+
         $article->body = $input['body'];
         $article->save();
-    
+
         return redirect()->route('articles.index');
     }
 
-    public function destroy(Article $article) {
+    public function destroy(Article $article)
+    {
+        // 컨트롤러 헬퍼를 사용
+        $this->authorize('delete', $article);
+
         $article->delete();
 
         return redirect()->route('articles.index');
